:refactor: reorganize utility function imports and move evaluate_annotation to private module
